+++ src/components/Sidebar/index.tsx
@@ -18,21 +18,26 @@ function Sidebar(props: SidebarProps) {
     },
     {
       id: 1,
+      name: 'Confide',
+      href: '/confide',
+    },
+    {
+      id: 2,
       name: 'About',
       href: '/about',
     },
     {
-      id: 2,
+      id: 3,
       name: 'Dev',
       href: '/dev',
     },
     {
-      id: 3,
+      id: 4,
       name: 'Book',
       href: '/book',
     },
     {
-      id: 4,
+      id: 5,
       name: 'Chore',
       href: '/chore',
     },
@@ -63,17 +68,15 @@ function Sidebar(props: SidebarProps) {
 
   return (
     <div className="tw-transition-all tw-shadow-xl tw-fixed tw-bg-gray-50 tw-w-80 tw-h-screen justify-between items-center tw-z-50 tw-top-0 tw-right-0 tw-border-l">
-      <div className='tw-h-16 tw-flex tw-justify-end tw-p-8 tw-items-center hover:tw-cursor-pointer' onClick={() => setIsOpenSidebar(!isOpenSidebar)}>
-        <Image
-          src={closeIcon}
-          alt='icon...'
-          width={20}
-          height={20}
-        />
+      <div
+        className="tw-h-16 tw-flex tw-justify-end tw-p-8 tw-items-center hover:tw-cursor-pointer"
+        onClick={() => setIsOpenSidebar(!isOpenSidebar)}
+      >
+        <Image src={closeIcon} alt="icon..." width={20} height={20} />
       </div>
       {renderMenuList}
     </div>
-  )
+  );
 }
 
 export default Sidebar;