+++ src/components/Tags/index.tsx
@@ -1,0 +1,13 @@
+import { TagsProps } from './types';
+
+function Tags(props?: TagsProps) {
+
+  return (
+    <div className=' tw-w-auto tw-rounded-md tw-relative'>
+      <div className='tw-w-full tw-bg-blue-300 tw-h-full rgb(blue) tw-absolute'></div>
+      <div className='tw-z-50'>tagne</div>
+    </div>
+  )
+}
+
+export default Tags;