+++ src/components/Tags/types.ts
@@ -1,0 +1,7 @@
+export interface TagsProps {
+  tagItemArr: {
+    color: string;
+    id: string;
+    name: string;
+  }[];
+}