+++ src/components/cards/CartItem/types.ts
@@ -1,0 +1,6 @@
+import { ArticleItemInterface } from '@/types/common';
+
+export interface CardItemProps {
+  data: ArticleItemInterface;
+  handleRedirectToDetail: (id: string) => void;
+}