+++ src/pages/book/index.tsx
@@ -1,7 +1,9 @@
-import MainLayout from '@/components/Layout/MainLayout';
-import BookCard from '@/components/cards/BookCard';
+import HomeLayout from '@/components/Layout/HomeLayout';
+import { useRouter } from 'next/router';
 import { getArticleList } from '@/services/api/getArticleList';
 import { ArticleListInterface } from '@/types/common';
+import CardItem from '@/components/cards/CartItem';
+import { generateBookDetailRouter } from '@/services/bussiness/router/generateRouter';
 
 export async function getServerSideProps() {
   const reqParam = {
@@ -26,15 +28,29 @@ export async function getServerSideProps() {
 
 function BookPage(props: any) {
   const data = props as ArticleListInterface;
+  const router = useRouter();
+
+  const handleClickViewDetailBook = (id: string) => {
+    const asPath = generateBookDetailRouter({ id: id })?.asPath;
+    if (asPath) {
+      router.push(asPath);
+    }
+  };
 
   return (
-    <div className="tw-grid tw-gap-4 tw-justify-items-center md:tw-grid-cols-2 tw-grid-cols-1 tw-mt-8 tw-mb-4">
-      {data?.results?.map((item, idx) => {
-        return <BookCard key={idx} data={item} />;
-      })}
+    <div className='tw-block'>
+      <div className='tw-mb-4'>
+        <span className='tw-text-blue-500 tw-font-semibold'> BOOK - bngoc tri thức :))  </span>
+      </div>
+
+      <div className="sm:tw-grid tw-block lg:tw-grid-cols-3  sm:tw-grid-cols-2 sm:tw-gap-4">
+        {data?.results?.map((item, idx) => {
+          return <CardItem key={idx} data={item} handleRedirectToDetail={handleClickViewDetailBook}/>;
+        })}
+      </div>
     </div>
   );
 }
 
-BookPage.Layout = MainLayout;
+BookPage.Layout = HomeLayout;
 export default BookPage;