+++ src/pages/chore/[id].tsx
@@ -100,7 +100,7 @@ function BookDetailPage(props: any) {
         </span>{' '}
         <br />
         <span className="tw-text-sm tw-text-slate-800">
-          Created at: {createDateConv ?? currDateInner}{' - '}
+        Created at: {createDateConv ?? currDateInner}{' - '}
           Updated at: {updatedAtConv ?? currDateInner}
         </span>{' '}
         <br />
@@ -118,6 +118,7 @@ function BookDetailPage(props: any) {
     </div>
   );
 }
+// Đi làm vui thấy bà
 
 BookDetailPage.Layout = EmptyLayout;
 export default BookDetailPage;