+++ src/pages/dev/index.tsx
@@ -1,7 +1,8 @@
-import MainLayout from '@/components/Layout/MainLayout';
-import DevCard from '@/components/cards/DevCard';
+import HomeLayout from '@/components/Layout/HomeLayout';
+import CardItem from '@/components/cards/CartItem';
 import { getArticleList } from '@/services/api/getArticleList';
 import { ArticleListInterface } from '@/types/common';
+import { useRouter } from 'next/router';
 
 export async function getServerSideProps() {
   const reqParam = {
@@ -25,17 +26,26 @@ export async function getServerSideProps() {
 }
 function DevPage(props: any) {
   const data = props as ArticleListInterface;
+  const router = useRouter();
+
+  const handleClickViewDetailDev = (id: string) => {
+    return router.push(`/dev/${id}`);
+  };
 
   return (
-    <div className=" tw-block">
-      <div className="tw-grid tw-gap-4 tw-justify-items-center md:tw-grid-cols-2 tw-grid-cols-1 tw-mt-8 tw-mb-4">
+    <div className='tw-block'>
+      <div className='tw-mb-4'>
+        <span className='tw-text-blue-500 tw-font-semibold'> DEVELOPER - Làm dev vui thấy bà  </span>
+      </div>
+
+      <div className="sm:tw-grid tw-block lg:tw-grid-cols-3  sm:tw-grid-cols-2 sm:tw-gap-4">
         {data?.results?.map((item, idx) => {
-          return <DevCard key={idx} data={item} />;
+          return <CardItem key={idx} data={item} handleRedirectToDetail={handleClickViewDetailDev} />;
         })}
       </div>
     </div>
   );
 }
 
-DevPage.Layout = MainLayout;
+DevPage.Layout = HomeLayout;
 export default DevPage;